+++ pmd-jsp/src/test/java/net/sourceforge/pmd/lang/jsp/rule/codestyle/CodeStyleRulesTest.java
@@ -1,0 +1,17 @@
+/**
+ * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
+ */
+
+package net.sourceforge.pmd.lang.jsp.rule.codestyle;
+
+import net.sourceforge.pmd.testframework.SimpleAggregatorTst;
+
+public class CodeStyleRulesTest extends SimpleAggregatorTst {
+
+    private static final String RULESET = "category/jsp/codestyle.xml";
+
+    @Override
+    public void setUp() {
+        addRule(RULESET, "DuplicateJspImports");
+    }
+}
