+++ pmd-jsp/src/test/java/net/sourceforge/pmd/lang/jsp/rule/errorprone/ErrorProneRulesTest.java
@@ -1,0 +1,17 @@
+/**
+ * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
+ */
+
+package net.sourceforge.pmd.lang.jsp.rule.errorprone;
+
+import net.sourceforge.pmd.testframework.SimpleAggregatorTst;
+
+public class ErrorProneRulesTest extends SimpleAggregatorTst {
+
+    private static final String RULESET = "category/jsp/errorprone.xml";
+
+    @Override
+    public void setUp() {
+        addRule(RULESET, "JspEncoding");
+    }
+}
