+++ pmd-jsp/src/test/resources/net/sourceforge/pmd/lang/jsp/rule/bestpractices/xml/DontNestJsfInJstlIteration.xml
