+++ pmd-jsp/src/test/resources/net/sourceforge/pmd/lang/jsp/rule/design/xml/NoScriptlets.xml
