+++ pmd-jsp/src/test/resources/net/sourceforge/pmd/lang/jsp/rule/errorprone/xml/JspEncoding.xml
